requirements.txt
fastapi==0.63.0
fr-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/fr_dep_news_trf-3.0.0/fr_dep_news_trf-3.0.0-py3-none-any.whl
sentence-transformers==1.1.0
spacy-transformers==1.0.2
emoji==1.2.0
torch==1.8.1
